requirements.txt
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
tensorflow>=2.12.0
PyYAML>=6.0
matplotlib>=3.5.0  # Optional: for visualization
tensorflow